fix(warnings): all pedantic warnings + norm errors

+++ src/list/at.c
@@ -6,7 +6,7 @@
 /*   By: smamalig <smamalig@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/26 20:20:15 by smamalig          #+#    #+#             */
-/*   Updated: 2025/03/19 15:24:28 by smamalig         ###   ########.fr       */
+/*   Updated: 2025/07/15 10:29:25 by smamalig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ t_list	*ft_list_at(t_list *list, ssize_t n)
 
 	if (n >= (ssize_t)size || n < -(ssize_t)size)
 		return (NULL);
-	index = n % size;
+	index = ((size_t)n) % size;
 	while (index-- && list)
 		list = list->next;
 	return (list);
