+++ src/str/modify/substr.c
@@ -6,7 +6,7 @@
 /*   By: smamalig <smamalig@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/10 16:34:37 by smamalig          #+#    #+#             */
-/*   Updated: 2025/04/14 14:16:14 by smamalig         ###   ########.fr       */
+/*   Updated: 2025/07/15 10:30:30 by smamalig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ char	*ft_substr(const char *s, size_t start, size_t size)
 	res = malloc(sizeof(char) * (size + 1));
 	if (!res)
 		return (NULL);
-	i = -1;
+	i = (size_t)-1;
 	while (++i < size)
 		res[i] = s[start + i];
 	res[i] = 0;
